refactor(capabilities): split the rpc server and test-echo actors into identity and runtime modules
Migrates RpcServerCapability + the test-echo actors off #[bridge] onto the
ADR-0122 identity/runtime split (the aether.fs template): ZST identities,
RpcServerState in a feature="runtime"-gated rpc/server/runtime.rs, handlers,
the #[fallback], and the deferred-echo #[handler(task)] all over
state: &mut Self::State. Thread shutdown is preserved — the accept and
per-connection reader threads capture init-local handles (Arc/mpsc/ids), not the
cap or state struct, so the unwire join is a field-home move.

Depended on the split-path macro completion landed earlier in this arc
(#[fallback] #2338, #[handler(task)] #2341).

Closes #2322.
