+++ crates/aether-capabilities/src/rpc/test_echo.rs
@@ -12,6 +12,20 @@
 
 use serde::{Deserialize, Serialize};
 
+// The actor halves are substrate-typed (ADR-0122 split). The whole module
+// is `#[cfg(test)]` (gated at its `mod` declaration in `rpc/mod.rs`) and
+// tests always carry `runtime`, so these resolve; the `#[actor]` macro
+// additionally gates the emitted `NativeActor` / `Dispatch` runtime impls
+// behind `feature = "runtime"`. The kind types above stay always-on so
+// their `Kind`-derived inventory submissions register for the test
+// substrate's registry walk.
+use crate::shared::contentgen::task_queue::TaskQueue;
+use aether_actor::actor;
+use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx, TaskDone};
+use aether_substrate::chassis::error::BootError;
+use std::thread;
+use std::time::Duration;
+
 /// Echo request kind — the test driver sends one of these; the echo
 /// actor replies with a [`TestEchoReply`] carrying the same `value`.
 #[derive(
@@ -53,37 +67,35 @@ pub struct TestEchoReply {
 /// matching [`TestEchoReply`]. The minimum viable receiver for
 /// exercising the RPC `Call → ReplyEvent → ReplyEnd` path without
 /// coupling a test to a production cap's semantics.
-#[aether_actor::bridge(singleton)]
-mod test_echo_actor {
-    use super::{TestEchoReply, TestEchoRequest};
-    use aether_actor::actor;
-    use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx};
-    use aether_substrate::chassis::error::BootError;
+/// `aether.rpc.test.echo` identity (ADR-0122 split). A ZST carrying only
+/// the addressing markers `#[actor]` emits always-on; the (empty) runtime
+/// state lives in `TestEchoActorState`.
+pub struct TestEchoActor;
 
-    pub struct TestEchoActor;
+/// Runtime state for [`TestEchoActor`]: a named empty stand-in (ADR-0122
+/// hard rule — never `()` / `Self`) for an actor that holds nothing.
+pub struct TestEchoActorState;
 
-    #[actor]
-    impl NativeActor for TestEchoActor {
-        type Config = ();
-        const NAMESPACE: &'static str = "aether.rpc.test.echo";
+#[actor(singleton)]
+impl NativeActor for TestEchoActor {
+    type State = TestEchoActorState;
+    type Config = ();
+    const NAMESPACE: &'static str = "aether.rpc.test.echo";
 
-        fn init((): (), _ctx: &mut NativeInitCtx<'_>) -> Result<Self, BootError> {
-            Ok(Self)
-        }
+    fn init((): (), _ctx: &mut NativeInitCtx<'_>) -> Result<TestEchoActorState, BootError> {
+        Ok(TestEchoActorState)
+    }
 
-        // Stateless echo handler — keeps `&mut self` to match the
-        // dispatch ABI (ADR-0033 / ADR-0038) even though the body
-        // doesn't touch component state.
-        #[allow(clippy::unused_self)]
-        #[handler]
-        fn on_echo(&mut self, _ctx: &mut NativeCtx<'_>, mail: TestEchoRequest) -> TestEchoReply {
-            TestEchoReply { value: mail.value }
-        }
+    /// Stateless echo handler — the empty state is unused.
+    #[handler]
+    fn on_echo(
+        _state: &mut Self::State,
+        _ctx: &mut NativeCtx<'_>,
+        mail: TestEchoRequest,
+    ) -> TestEchoReply {
+        TestEchoReply { value: mail.value }
     }
 }
-
-// `TestEchoActor` is re-exported at this module's root by the
-// `#[bridge]` macro itself — no explicit `pub use` needed.
 
 /// Deferred-echo request — like [`TestEchoRequest`] but the actor
 /// answers it through the ADR-0093 hold-until-resolve dispatch
@@ -137,60 +149,59 @@ pub struct DeferredEchoReply {
 /// -> re-reply). The whole point is that the reply happens *after* the
 /// handler returns, so the framework-held settlement hold must keep the
 /// chain open across the gap.
-#[aether_actor::bridge(singleton)]
-mod deferred_echo_actor {
-    use super::{DeferredEchoReply, DeferredEchoRequest};
-    use crate::shared::contentgen::task_queue::TaskQueue;
-    use aether_actor::actor;
-    use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx, TaskDone};
-    use aether_substrate::chassis::error::BootError;
-    use std::thread;
-    use std::time::Duration;
+/// `aether.rpc.test.deferred_echo` identity (ADR-0122 split). A ZST
+/// carrying the addressing markers; its runtime state — the
+/// `TaskQueue` backing the off-thread dispatch — lives in
+/// `DeferredEchoActorState`.
+pub struct DeferredEchoActor;
 
-    pub struct DeferredEchoActor {
-        tasks: TaskQueue,
-    }
-
-    #[actor]
-    impl NativeActor for DeferredEchoActor {
-        type Config = ();
-        const NAMESPACE: &'static str = "aether.rpc.test.deferred_echo";
-
-        fn init((): (), _ctx: &mut NativeInitCtx<'_>) -> Result<Self, BootError> {
-            Ok(Self {
-                tasks: TaskQueue::new(4),
-            })
-        }
-
-        /// Submit the echo off-thread via the ADR-0093 dispatch primitive.
-        /// The worker sleeps briefly so the handler reliably returns
-        /// (queuing its `Finished`) before the reply lands — the window the
-        /// bug used to settle in. The framework-held `SettlementHold` keeps
-        /// the chain open until the deferred re-reply.
-        #[handler]
-        fn on_deferred_echo(&mut self, ctx: &mut NativeCtx<'_>, mail: DeferredEchoRequest) {
-            let value = mail.value;
-            self.tasks.submit(ctx, move || {
-                // Brief blocking work standing in for a provider call.
-                thread::sleep(Duration::from_millis(50));
-                DeferredEchoReply { value }
-            });
-        }
-
-        /// ADR-0093 completion: re-reply to the original caller (drops the
-        /// hold after the reply — `Sent` precedes `Release`), then free the
-        /// in-flight slot.
-        #[handler(task)]
-        fn on_deferred_echo_done(
-            &mut self,
-            ctx: &mut NativeCtx<'_>,
-            done: TaskDone<DeferredEchoReply>,
-        ) {
-            done.resolve(ctx);
-            self.tasks.on_complete(ctx);
-        }
-    }
+/// Runtime state for [`DeferredEchoActor`]: the ADR-0093 hold-until-resolve
+/// task queue the deferred handler submits onto.
+pub struct DeferredEchoActorState {
+    tasks: TaskQueue,
 }
 
-// `DeferredEchoActor` is re-exported at this module's root by the
-// `#[bridge]` macro.
+#[actor(singleton)]
+impl NativeActor for DeferredEchoActor {
+    type State = DeferredEchoActorState;
+    type Config = ();
+    const NAMESPACE: &'static str = "aether.rpc.test.deferred_echo";
+
+    fn init((): (), _ctx: &mut NativeInitCtx<'_>) -> Result<DeferredEchoActorState, BootError> {
+        Ok(DeferredEchoActorState {
+            tasks: TaskQueue::new(4),
+        })
+    }
+
+    /// Submit the echo off-thread via the ADR-0093 dispatch primitive.
+    /// The worker sleeps briefly so the handler reliably returns
+    /// (queuing its `Finished`) before the reply lands — the window the
+    /// bug used to settle in. The framework-held `SettlementHold` keeps
+    /// the chain open until the deferred re-reply.
+    #[handler]
+    fn on_deferred_echo(
+        state: &mut Self::State,
+        ctx: &mut NativeCtx<'_>,
+        mail: DeferredEchoRequest,
+    ) {
+        let value = mail.value;
+        state.tasks.submit(ctx, move || {
+            // Brief blocking work standing in for a provider call.
+            thread::sleep(Duration::from_millis(50));
+            DeferredEchoReply { value }
+        });
+    }
+
+    /// ADR-0093 completion: re-reply to the original caller (drops the
+    /// hold after the reply — `Sent` precedes `Release`), then free the
+    /// in-flight slot.
+    #[handler(task)]
+    fn on_deferred_echo_done(
+        state: &mut Self::State,
+        ctx: &mut NativeCtx<'_>,
+        done: TaskDone<DeferredEchoReply>,
+    ) {
+        done.resolve(ctx);
+        state.tasks.on_complete(ctx);
+    }
+}
